Only restart 'salt-api' is salt-api is installed.
Salt is used by sumaform for provisioning, so `salt` package is always installed in every deployed machine but not the `salt-api` package.

This PR fixes the "HACK" to restart the `salt-api` in order to prevent errors during provisioning if `salt-api` doesn't exist.

Also prevent an error trying to run `systemctl restart salt-api` if running OS uses SysV.