Refactoring: use immutable images
This is a change in the overall design. Until this point, images were supposed
to be kept up-to-date (eg. by being built on OBS and refreshed by clients from
time to time).

Now we want to create images once and then forget about them - all updates will
always come from packages and be controlled by Salt. This has the advantage of
controlling the Salt version using during deployment, which is important in some
scenarios.

sumaform will still work with old images.