Repo reorganization
We want to control the version of Salt we use for provisioning. In order to do that,
all VMs must have correct base repositories and, in case of SLES, the SUSE Manager
Tools channels that ship Salt.

We use the existing default state to contain this, and make all other states depend on
it.