Force package upgrade as a separate step
As per the new design, we can't rely on images having updated packages,
so we update them as part of the default state.

This might mean Salt itself is updated, and this is known to cause errors
if any output formatter is used and if anything is executed after the update.

So we piggy-back package upgrading to the terraform-resource state application,
which already happened as a separate Salt call.

This way the second Salt call will run with the desired Salt version in a reliable
way.