Refactoring: isolate salt and zypper upgrade from all other changes
This patch splits functionality in terraform-resource putting it in the
default state, in such a way it makes sense even outside of the
Terraform context.

Moreover it changes the first provisioning Salt call to only apply the
one salt state ID - the one that upgrades Salt and any other essential
package, leaving all the rest to the second call to minimize
interference.