Taint an image resource
It happens sometime you want to refresh your images, so you want to `taint` them to get a new downloaded image.
The syntax is the same when tainting any other resource, but the name of the module and the number of the volume are slightly different. I think it can be useful to have it documented here to get it quickly done.