Revert policy to have the FQDN in /etc/hostname
It is debated whether to have the "short name" or the "fully qualified
domain name" as contents of /etc/hostname, as some software seems to be
sensible to this. Debian systems typically use the "short name", while
Red Hat derivatives take the opposite choice.

Main disadvantage of the "short name" in /etc/hostname is that
`hostname -f` does not return the correct value in absence of a working
DNS, which is why the policy is reverted here.