Add virthost module
For SUSE Manager / Uyuni to handle virtualized virtual hosts we need to
lure salt and get it to think that VM is physical. There are two things
that are checked:

  * systemd-detect-virt (checks CPUID and DMI values)
  * dmidecode output should not have any QEMU manufacturer

There are two possible ways to fake the dmidecode output. One is using
the <smbios> element in the libvirt config and the -smbios qemu argument
to cover the missing types (CPU: 4 and Memory: 17). The other option is
to replace the dmidecode tool by a script calling dmidecode with a
predefined binary file. This commit opts for the second solution since
it doesn't enfore the use of qemu as hypervisor for sumaform and is
easier to deploy.

Faking the CPUID is way more complex if possible at all. Then this
commit just replaces the systemd-detect-virt by a script returning the
output we need.

The virtual host is initialized with a default pool and a default
network to ease testing with it.

Matrix of nesting support:

  host  | guest | Status
  ------------------------
  KVM   | KVM   | OK
  KVM   | Xen   | Doesn't work
  Xen   | KVM   | QEMU only, no KVM acceleration
  Xen   | Xen   | Doesn't work (at least packaging issues)