suse_manager_proxy: do not fail when resource tuple is empty
During or after destroy, outputs of resources with a count attribute
may contain empty tuples, and that will fail at destroy or apply time.

This patch makes sure to avoid indexing empty tuples. Note that previous
code checked the non-emptiness of tuples indirectly, depending on input
variables, but this fails to catch all cases - a 0 cound might either
depend on input variables or from the destruction of earlier resources.
With this patch we catch all cases.