Add buildhost VM
The idea is to have one VM for the 'salt minion' and another VM
for the 'Docker and Kiwi build host' in the Cucumber test suite.

This change adds flexibility to the test suite; by decoupling the 'salt minion'
from the build host we a are able to test more minion versions (not just
those with a build host functionality available).

Another advantage is a reduction in the number of clean up scenarios
(the slow readd entitlements).
See spacewalk#10680

- modules/controller/variables.tf
  Add buildhost configuration.

- modules/cucumber_testsuite/main.tf
  Add min-build module section.
  Remove avahi_reflector from the normal minion; keep it only for the min-build.

- salt/controller/bashrc:
  Export environment variable with the build host hostname.

- modules/controller/main.tf
  Read build host configuration.

- README_TESTING.md (Advanced deployments):
  Include the new VM in the example.