Use a more stable image for the controller image
opensuse 15.2 images get updated very frequently. This goes to the
openSUSE mirrors. From time to time, images are not available in the
mirrors. This causes a failure when running sumaform.

Let's use a more stable image, that is not updated and, so it is available
most of the time, plus an image that is specific for the
controller, which its build is being disabled.

Moreover, this will make a run reproduceable. If we use an image that is
continuously being updated, runs are not reproduceable.

Signed-off-by: Jordi Massaguer Pla <jmassaguerpla@suse.de>