Fix installation of containerized proxy systemd services
This package lives in the client tools repo, but to reduce the impact on
the testsuite we decided to reuse the proxy machine for containerized
proxy tests.

To avoid manual download and install of the RPM as this leads to issues,
temporarily add the client tools repo to the proxy.