Fix early avahi installation on controller
Previously, since the controller always runs on Leap, the custom repo
was not added in the early stage of the controller's setup. Because
of this, the installation of avahi-lang would always fail if avahi
is enabled in main.tf. This usually wasn't noticed, because a later
step will also attempt to install this package, which would then
succeed. However, there might be a need for avahi being available
during the early stages.