Refactor salt_classic_package.sls
A lot of repetition is avoided by using jinja features. This is a test, if
we like using it we can expand it to more of sumaform.

I wanted to use less `cmd.run` and hit a known bug for salt-call in
MicroOS, where `--local` is not passed. I left the intended code in the
file, commented out, until we fix Salt.